Enable lots of warnings by default.
More changes to zap warnings. cmake/setup_target.cmake was lifted from
my icubaby project. It enables lots on warnings on the various
compilers, supports coverage analysis, sanitizers, hardening options,
and so on.

+++ include/utils.h
@@ -96,7 +96,7 @@ enum : std::uint32_t {
   ENDOFFILE = 0x021,
 };
 
-enum : std::uint8_t {
+enum : std::uint32_t {
   FUNCTIONBLOCK = 0x010,
   FUNCTIONBLOCK_INFO_NOTFICATION = 0x011,
   FUNCTIONBLOCK_NAME_NOTIFICATION = 0x012,
